o.y.cmake.managedbuilder: resolve absolute path to cmake executable
The cmake project appears to always use host cmake unless an absolute
path to the executable in SDK or build directory is specified in project
settings, this is likely caused by a misunderstanding in how the Java
ProcessBuilder works, where setting the PATH environment variable to
include path to SDK or build directory would not help to resolve the
location of the cmake executable.

With this fix, the cmake process would be executed using absolute path
resolved from the latest value PATH environment variable, and in the
event that SDK or build directory paths are not set, this would fallback
to use the original PATH value to resolve host cmake.

Signed-off-by: Chin Huat Ang <chin.huat.ang@intel.com>