org.yocto.cmake.core: don't resolve cmake executable path
The current approach for resolving cmake executable's absolute path does
not work with ContainerCommandLauncher as it is not possible to use
java.io.File to test the existence of the executable. Instead we relax
or delegate to the underlying command launcher to determined the best
way to resolve the executable's absolute path. Alternatively, this can
also be done in CMakeMakefileGenerator subclasses.

Signed-off-by: Chin Huat Ang <chin.huat.ang@intel.com>