setup.sh: use Yocto Project full eclipse mirror
We now have a full mirror of Eclipse, so there is no
longer any reason to go to upstream for UPDATE_SITE.

For the other fetches that we were doing from the old,
manually updated eclipse mirror, change to use the
full mirror.

Should help with [YOCTO #13050], but not a necessarily the
complete fix.

Signed-off-by: Tim Orling <timothy.t.orling@linux.intel.com>